Fixed effects not being added

+++ src/io/anuke/ucore/entities/Entity.java
@@ -3,11 +3,11 @@ package io.anuke.ucore.entities;
 import com.badlogic.gdx.math.Vector2;
 
 public abstract class Entity{
-	private static long lastid;
+	private static int lastid;
 	protected static Vector2 vector = new Vector2();
 	public static float delta;
 	
-	public final long id;
+	public final int id;
 	public float x,y;
 	
 	public void update(){}
